GC+: Fix list objects from store
